requirements.txt
asgiref==3.10.0
cffi==2.0.0
cryptography==46.0.3
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
mssql-django==1.6
mysqlclient==2.2.7
oracledb==3.4.0
psycopg2-binary==2.9.11
pycparser==2.23
PyMySQL==1.1.2
pyodbc==5.3.0
python-decouple==3.8
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.15.0
